use int for index and size_t for sizes for consquently

+++ src/tinyspline.h
@@ -43,8 +43,8 @@ typedef struct
 {
     float u;           // <- the knot u
     int k;             // <- the index [u_k, u_k+1)
-    int s;             // <- the multiplicity of u_k
-    int h;             // <- how many times u must be inserted
+    size_t s;          // <- the multiplicity of u_k
+    size_t h;          // <- how many times u must be inserted
     size_t deg;        // <- degree of b-spline bassis function
     size_t dim;        // <- dimension of a control point
     size_t n_affected; // <- number of affected control points
